requirements.txt
fastapi
uvicorn
pydantic
google-genai
a2wsgi